requirements/dev.txt
-r ./prod.txt

atomicwrites==1.2.1
attrs==18.2.0
more-itertools==4.3.0
pluggy==0.7.1
pytest==3.8.2
pytest-django==3.4.3
six==1.11.0
flake8==3.5.0
flake8-html==0.4.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
coverage==4.5.1
unittest-xml-reporting==2.2.0
Faker==1.0.0
factory-boy==2.11.1
text-unidecode==1.2
django-extensions==2.1.6